requirements.txt
django==3.0.6
django-cors-header==3.3.0
djangorestframework==3.11.0
